The following four functions all include unit testers.
Adding heuristic_eval_function to predict the winner if not at the end
of the game.
Adding at_game_end function to check if the end of the game is reached.
Adding a determine_winner function to determine the winner at the end of
the game.
Adding a check_and_print_victory_conditions module to print the winner.

These changes also necessitated changes to the driver module to support
them.